web/requirements.txt
django==4.0
psycopg2-binary==2.9.2
gunicorn==20.0.4
djangorestframework==3.13.1
